requirements.txt
folium==0.20.0
googlemaps==4.10.0
pandas==2.3.2
pyppeteer==2.0.0
python-dotenv==1.1.1
python_docx==1.2.0
streamlit==1.49.1
altair==4.0.0
PyPDF2==3.0.1
